Splits version bumping and publishing into separate scripts
Refactors the `Publish-MyModule.ps1` script into two separate scripts:
`Set-Version.ps1` and a future `Publish-MyModule.ps1`.

`Set-Version.ps1` is responsible for updating the module
version in the module manifest file.  It takes a path to the
module and an optional `IncrementVersion` parameter to specify
which part of the version to increment.

This change separates concerns and prepares for a future
publishing script which will be used for actually publishing to
the PSGallery.