Changes output to debug messages
Updates Write-Output to Write-Debug in Set-EntraIdGroup and
Set-EntraIdUser to reduce console noise during normal operations.

This change makes the scripts less verbose during normal operation
and ensures that informational messages are only displayed when
debugging is enabled, providing a cleaner output when running
the scripts in production or without the need for detailed logging.